requirements.txt
fastapi
uvicorn
boto3
python-multipart
python-dotenv
python-telegram-bot
requests
pytest
python-jose[cryptography]


